Move qt_available and require_pyqt4_api2 to core
