+++ vistrails/core/requirements.py
@@ -38,6 +38,7 @@
 runtime components such as binaries, libraries, other python modules, etc."""
 from __future__ import division
 
+import os
 import sys
 
 import vistrails.core.bundles.installbundle
@@ -95,3 +96,42 @@ class MissingRequirement(Exception):
     def __str__(self):
         return "Missing Requirement: %s" % self.requirement
 
+##############################################################################
+
+def setNewPyQtAPI():
+    import sip
+    # We now use the new PyQt API - IPython needs it
+    sip.setapi('QString', 2)
+    sip.setapi('QVariant', 2)
+
+
+def qt_available():
+    try:
+        require_python_module('sip')
+        setNewPyQtAPI()
+        require_python_module('PyQt4.QtGui')
+        require_python_module('PyQt4.QtOpenGL')
+    except MissingRequirement:
+        return False
+    else:
+        return True
+
+
+def require_pyqt4_api2():
+    # Forces the use of PyQt4 (avoid PySide even if installed)
+    # This is necessary at least for IPython
+    if os.environ.get('QT_API', None) not in (None, 'pyqt'):
+        sys.stderr.write("Warning: QT_API was set to %r, changing to 'pyqt'\n" %
+                         os.environ['QT_API'])
+    os.environ['QT_API'] = 'pyqt'
+
+    if not qt_available():
+        from vistrails.gui.bundles.installbundle import install
+        r = install({
+            'linux-debian': ['python-qt4', 'python-qt4-gl', 'python-qt4-sql'],
+            'linux-ubuntu': ['python-qt4', 'python-qt4-gl', 'python-qt4-sql'],
+            'linux-fedora': ['PyQt4'],
+            'pip': ['PyQt<5.0']})
+        if not r:
+            raise MissingRequirement('PyQt4')
+        setNewPyQtAPI()
